refactoring, error passing and tests
