+++ src/Config/ConfigurationStore.php
@@ -7,7 +7,7 @@ use Eppo\Cache\CacheType;
 use Eppo\Cache\NamespaceCache;
 use Eppo\DTO\Bandit\Bandit;
 use Eppo\DTO\Flag;
-use Eppo\Exception\EppoClientException;
+use Eppo\Exception\InvalidConfigurationException;
 use Psr\SimpleCache\CacheInterface;
 use Psr\SimpleCache\InvalidArgumentException;
 
@@ -32,6 +32,11 @@ class ConfigurationStore implements IConfigurationStore
         $this->metadataCache = new NamespaceCache(CacheType::META, $cache);
     }
 
+    /**
+     * @param string $key
+     * @return Flag|null
+     * @throws InvalidConfigurationException
+     */
     public function getFlag(string $key): ?Flag
     {
         try {
@@ -44,28 +49,30 @@ class ConfigurationStore implements IConfigurationStore
             return $inflated === false ? null : $inflated;
         } catch (InvalidArgumentException $e) {
             // Simple cache throws exceptions when a keystring is not a legal value (characters {}()/@: are illegal)
-            syslog(LOG_WARNING, "[EPPO SDK] Illegal flag key ${key}: " . $e->getMessage());
-            return null;
+            throw new InvalidConfigurationException("Illegal flag key ${key}: " . $e->getMessage(), $e->getCode(), $e);
         }
     }
 
+    /**
+     * @throws InvalidConfigurationException
+     */
     private function setFlag(Flag $flag): void
     {
         try {
             $this->flagCache->set($flag->key, serialize($flag));
         } catch (InvalidArgumentException $e) {
             $key = $flag->key;
-
             // Simple cache throws exceptions when a keystring is not a legal value (characters {}()/@: are illegal)
-            syslog(LOG_WARNING, "[EPPO SDK] Illegal flag key ${key}: " . $e->getMessage());
+            throw new InvalidConfigurationException("Illegal flag key ${key}: " . $e->getMessage(), $e->getCode(), $e);
         }
     }
 
     /**
      * @param array $flags
-     * @param Bandit[] $bandits
+     * @param array $bandits
      * @param BanditVariationIndexer|null $banditVariations
-     * @throws EppoClientException
+     * @return void
+     * @throws InvalidConfigurationException
      */
     public function setConfigurations(
         array $flags,
@@ -82,10 +89,13 @@ class ConfigurationStore implements IConfigurationStore
             $this->setBandits($bandits);
             $this->metadataCache->set(self::BANDIT_VARIATION_KEY, serialize($banditVariations));
         } catch (InvalidArgumentException $e) {
-            throw EppoClientException::from($e);
+            throw InvalidConfigurationException::from($e);
         }
     }
 
+    /**
+     * @throws InvalidConfigurationException
+     */
     private function setFlags(array $flags): void
     {
         foreach ($flags as $flag) {
@@ -107,7 +117,8 @@ class ConfigurationStore implements IConfigurationStore
     }
 
     /**
-     * @throws EppoClientException
+     * @return BanditVariationIndexer
+     * @throws InvalidConfigurationException
      */
     public function getBanditVariations(): BanditVariationIndexer
     {
@@ -115,15 +126,19 @@ class ConfigurationStore implements IConfigurationStore
             return unserialize($this->metadataCache->get(self::BANDIT_VARIATION_KEY));
         } catch (InvalidArgumentException $e) {
             // We know that the key does not contain illegal characters, so we should not end up here.
-            throw EppoClientException::From($e);
+            throw InvalidConfigurationException::From($e);
         }
     }
 
+    /**
+     * @param Bandit[] $bandits
+     * @return void
+     */
     private function setBandits(array $bandits): void
     {
         foreach ($bandits as $bandit) {
             try {
-                $this->banditCache->set($bandit->key, serialize($bandit));
+                $this->banditCache->set($bandit->banditKey, serialize($bandit));
             } catch (InvalidArgumentException $e) {
                 $message = $e->getMessage();
                 syslog(LOG_WARNING, "[Eppo SDK]: Error \"{$message}\" encountered while setting bandit");
@@ -131,15 +146,23 @@ class ConfigurationStore implements IConfigurationStore
         }
     }
 
+
     /**
-     * @throws EppoClientException
+     * @param string $banditKey
+     * @return Bandit|null
+     * @throws InvalidConfigurationException
      */
-    public function getBandit(string $banditKey): array
+    public function getBandit(string $banditKey): ?Bandit
     {
         try {
             return $this->banditCache->get($banditKey);
         } catch (InvalidArgumentException $e) {
-            throw EppoClientException::From($e);
+            // Simple cache throws exceptions when a keystring is not a legal value (characters {}()/@: are illegal)
+            throw new InvalidConfigurationException(
+                "Illegal bandit key ${$banditKey}: " . $e->getMessage(),
+                $e->getCode(),
+                $e
+            );
         }
     }
 }
