+++ src/DTO/Bandit/BanditVariation.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Eppo\DTO\Bandit;
 
 class BanditVariation
