+++ src/Exception/InvalidConfigurationException.php
@@ -11,4 +11,9 @@ class InvalidConfigurationException extends Exception
     {
         parent::__construct($message, $code, $previous);
     }
+
+    public static function from(Throwable $previous = null, int $code = 0): self
+    {
+        return new self($previous->getMessage(), $code, $previous);
+    }
 }
