+++ src/Flags/IFlags.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Eppo;
+namespace Eppo\Flags;
 
 use Eppo\DTO\Flag;
+use Eppo\Exception\InvalidConfigurationException;
 
 /**
  * A collection of Flags indexed by key.
@@ -13,6 +14,7 @@ interface IFlags
      * Gets a flag from the collection if it exists.
      * @param string $key
      * @return ?Flag
+     * @throws InvalidConfigurationException
      */
     public function getFlag(string $key): ?Flag;
 }
