+++ tests/DTO/Bandit/AttributeSetTest.php
